move RecycleView to TimelineFragment
