+++ app/src/main/res/layout/fragment_tracktime.xml
@@ -1,46 +1,27 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
     <Spinner
         android:id="@+id/projectSpinner"
-        android:layout_width="230dp"
-        android:layout_height="32dp"
-        android:layout_marginTop="4dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.823"
+        android:layout_width="193dp"
+        android:layout_height="36dp"
+        android:layout_marginTop="66dp"
+        app:layout_constraintEnd_toStartOf="@+id/startTask"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/endTask" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycler"
-        android:layout_width="329dp"
-        android:layout_height="213dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="120dp"
-        android:layout_marginEnd="32dp"
-        android:scrollbars="vertical"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.484"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/taskNameShow"
-        tools:context="com.github.studenttimetracker.MainActivity">
-
-    </androidx.recyclerview.widget.RecyclerView>
+        app:layout_constraintTop_toTopOf="@+id/taskNameInput" />
 
     <Chronometer
         android:id="@+id/myChronometer"
-        android:layout_width="245dp"
-        android:layout_height="47dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="12dp"
-        android:textSize="25sp"
-        app:layout_constraintEnd_toStartOf="@+id/endTask"
-        app:layout_constraintHorizontal_bias="0.0"
+        android:layout_width="387dp"
+        android:layout_height="75dp"
+        android:layout_marginTop="36dp"
+        android:gravity="center"
+        android:textSize="50sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.562"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -48,68 +29,70 @@
         android:id="@+id/endTask"
         android:layout_width="88dp"
         android:layout_height="48dp"
-        android:layout_marginEnd="32dp"
+        android:layout_marginTop="16dp"
         android:text="@string/stop_task"
-        app:layout_constraintBaseline_toBaselineOf="@+id/myChronometer"
-        app:layout_constraintEnd_toEndOf="parent" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.498"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/taskNameShow" />
 
     <TextView
         android:id="@+id/taskNameShow"
-        android:layout_width="183dp"
-        android:layout_height="21dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="4dp"
+        android:layout_width="178dp"
+        android:layout_height="38dp"
+        android:layout_marginBottom="24dp"
         android:fontFamily="@font/roboto_medium"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/myChronometer" />
+        android:gravity="center"
+        android:textSize="30sp"
+        app:layout_constraintBottom_toTopOf="@+id/projectSpinner"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.497"
+        app:layout_constraintStart_toStartOf="parent" />
 
     <EditText
         android:id="@+id/taskNameInput"
-        android:layout_width="197dp"
-        android:layout_height="46dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="12dp"
+        android:layout_width="352dp"
+        android:layout_height="44dp"
+        android:layout_marginStart="20dp"
+        android:layout_marginTop="4dp"
+        android:layout_marginEnd="20dp"
         android:autofillHints=""
         android:ems="10"
         android:hint="@string/task_name_hint"
         android:inputType="textPersonName"
-        app:layout_constraintEnd_toStartOf="@+id/startTask"
-        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toBottomOf="@+id/myChronometer" />
 
     <Button
         android:id="@+id/startTask"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="32dp"
+        android:layout_width="109dp"
+        android:layout_height="40dp"
+        android:layout_marginEnd="20dp"
         android:text="@string/start_task"
-        app:layout_constraintBaseline_toBaselineOf="@+id/taskNameInput"
+        app:layout_constraintBaseline_toBaselineOf="@+id/endTask"
         app:layout_constraintEnd_toEndOf="parent" />
 
     <Button
         android:id="@+id/addProject"
         android:layout_width="84dp"
         android:layout_height="38dp"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="32dp"
         android:text="@string/add_project"
+        app:layout_constraintBaseline_toBaselineOf="@+id/projectName"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/projectSpinner" />
+        app:layout_constraintStart_toEndOf="@+id/projectName" />
 
     <EditText
         android:id="@+id/projectName"
         android:layout_width="194dp"
         android:layout_height="44dp"
-        android:layout_marginEnd="12dp"
-        android:ems="10"
-        android:inputType="textPersonName"
+        android:layout_marginStart="20dp"
+        android:layout_marginTop="76dp"
         android:autofillHints=""
+        android:ems="10"
         android:hint="@string/enter_project_hint"
-        app:layout_constraintBaseline_toBaselineOf="@+id/addProject"
-        app:layout_constraintEnd_toStartOf="@+id/addProject"
-        app:layout_constraintHorizontal_bias="0.842"
-        app:layout_constraintStart_toStartOf="parent" />
+        android:inputType="textPersonName"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/endTask" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
